win: show console in command line mode
Fucking Windows doesn't allow application to be both
console and gui at the same. No way to warkoround this.
So I use regular gui application and open new console
window for command line mode. It is mean that when torrent-file editor
is started from console with --help or --to-json or --from-json
arguments it will always open console window to output text.
Unsure but I can't just put text in old cmd.exe window.